+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Builtin Apps:
 
 Board Support:
 - Fri3d 2024/2026: correct IR TX pin for SAO IO13/IO21 (not IO10 badge link) so Mini Blaster / Noisy Cricket works with IR Remote app
+- Fri3d 2026: add support for headset-only audio output (without simulteneous communicator output)
 
 Drivers:
 - ST7789 display controller: tune up VCOMS, VRHS for improved contrast; raise frame rate to 90 Hz to prevent tearing effect; update positive and negative gamma curves for better colors